feat: Changes to config displaying and header structure
