+++ frontend/src/i18n/messages.js
@@ -83,6 +83,7 @@ export const messages = {
         logout: 'Logout',
         identityTitle: 'Visual identity',
         identityHint: 'Reserved space for the pena or league crest.',
+        appearanceTitle: 'Appearance & language',
         identityPlaceholder: 'Crest',
         profileImageLabel: 'Profile image',
         imageEmpty: 'No image selected',
@@ -919,6 +920,7 @@ export const messages = {
         logout: 'Cerrar sesión',
         identityTitle: 'Identidad visual',
         identityHint: 'Espacio reservado para el escudo de la peña o de la liga.',
+        appearanceTitle: 'Apariencia e idioma',
         identityPlaceholder: 'Escudo',
         profileImageLabel: 'Imagen de perfil',
         imageEmpty: 'No hay imagen seleccionada',
